remove javadb oci download
