+++ pkg/javadb/download_no_db.go
@@ -1,0 +1,13 @@
+//go:build trivy_no_javadb
+
+package javadb
+
+import (
+	"context"
+
+	"golang.org/x/xerrors"
+)
+
+func (u *Updater) downloadDB(ctx context.Context) error {
+	return xerrors.Errorf("download Java DB unavailable")
+}
